# docs/swagger.yaml
basePath: /
definitions:
...
  auth.TokenResp:
    properties:
      accessToken:
        type: string
    type: object
  auth.UpdateProfileReq:
    properties:
      confirmNewPassword:
        maxLength: 64
        minLength: 8
        type: string
      name:
        type: string
      newPassword:
        maxLength: 64
        minLength: 8
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    type: object
...
  web.ErrResp:
    properties:
      error:
        type: string
      fields:
        items:
          $ref: '#/definitions/web.FieldError'
        type: array
    type: object
  web.FieldError:
    properties:
      error:
        type: string
      field:
        type: string
    type: object
host: localhost:8001
info:
  contact: {}
  description: API for game library auth service
  termsOfService: http://swagger.io/terms/
  title: Game library auth API
  version: "0.4"
paths:
  /account:
    delete:
      description: Deletes a user account
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted account
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrResp'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrResp'
      summary: Delete user account
      tags:
      - auth
    patch:
      consumes:
      - application/json
      description: Updates the profile information of a user
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update profile parameters
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/auth.UpdateProfileReq'
      produces:
      - application/json
      responses:
        "200":
          description: Returns new access token
          schema:
            $ref: '#/definitions/auth.TokenResp'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrResp'
        "401":
          description: Invalid password or token
          schema:
            $ref: '#/definitions/web.ErrResp'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/web.ErrResp'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrResp'
      security:
      - Bearer: []
      summary: Update user profile
      tags:
      - auth
...
schemes:
- http
swagger: "2.0"